Fixed a bug related to the appearance of a large number of digits after the dot.

+++ app/src/main/java/com/example/converter/Unit.kt
@@ -5,6 +5,6 @@ import java.math.BigDecimal
 
 data class Unit(
     @StringRes val Name: Int,
-    var multiplier: BigDecimal,
+    var multiplier: Int,
     val group: UnitGroup
 )
